move BudgetPositionsFields to fields folder
